% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_circle.R
\name{lattice_circle}
\alias{lattice_circle}
\title{Generate a Lattice for Circle}
\usage{
lattice_circle(n = 5, display_plot = TRUE)
}
\arguments{
\item{n}{A positive integer specifying the number of nodes in the graph.}

\item{display_plot}{Logical. If \code{TRUE}, plot the resulting lattice using \code{plot.network()} from the \strong{network} package.}
}
\value{
A list with the following components:
\describe{
\item{\code{adj}}{A symmetric adjacency matrix representing the lattice for circle.}
\item{\code{coord}}{A data frame of node coordinates for plotting.}
\item{\code{net}}{A \code{network} object from the \strong{network} package representing the lattice.}
}
}
\description{
Construct an adjacency matrix for a circle (i.e., cycle graph) with \code{n} nodes.
}
\examples{
lat <- lattice_circle(n = 5, display_plot = FALSE)
lat$adj
plot(lat$net, coord = lat$coord, jitter = FALSE, displaylabels = TRUE)
}
