% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_complete.R
\name{lattice_complete}
\alias{lattice_complete}
\title{Generate a Lattice for Full Connection}
\usage{
lattice_complete(n = 5, display_plot = TRUE)
}
\arguments{
\item{n}{A positive integer specifying the number of nodes in the graph.}

\item{display_plot}{Logical. If \code{TRUE}, the lattice is plotted using \code{plot.network()} from the \strong{network} package.}
}
\value{
A list with the following components:
\describe{
\item{\code{adj}}{A symmetric adjacency matrix representing the complete graph.}
\item{\code{coord}}{A data frame of node coordinates arranged in a circle.}
\item{\code{net}}{A \code{network} object from the \strong{network} package representing the graph.}
}
}
\description{
Construct an adjacency matrix for a fully connected lattice (i.e., complete graph) with \code{n} nodes.
}
\examples{
lat <- lattice_complete(n = 5, display_plot = FALSE)
lat$adj
plot(lat$net, coord = lat$coord, jitter = FALSE, displaylabels = TRUE)
}
