% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_cube.R
\name{lattice_cube}
\alias{lattice_cube}
\title{Generate a Lattice for Cube}
\usage{
lattice_cube(nx = 4, ny = 4, nz = 4, torus = TRUE, display_plot = TRUE)
}
\arguments{
\item{nx}{Integer. Number of nodes along the x-axis. Must be 2 or greater.}

\item{ny}{Integer. Number of nodes along the y-axis. Must be 2 or greater.}

\item{nz}{Integer. Number of nodes along the z-axis. Must be 2 or greater.}

\item{torus}{Logical. If \code{TRUE}, wrap the lattice around each axis to form a 3D torus.}

\item{display_plot}{Logical. If \code{TRUE}, plot the resulting lattice using \code{plot.network()} from the \strong{network} package.}
}
\value{
A list with the following components:
\describe{
\item{\code{adj}}{A symmetric adjacency matrix representing the lattice for cube.}
\item{\code{coord}}{A data frame of node coordinates for plotting.}
\item{\code{net}}{A \code{network} object from the \strong{network} package representing the lattice.}
}
}
\description{
Construct an adjacency matrix for a cubic shape lattice with \eqn{n_x \times n_y \times n_z} nodes, optionally wrapped as a torus.
}
\examples{
lat <- lattice_cube(nx = 4, ny = 4, nz = 4, torus = TRUE, display_plot = FALSE)
lat$adj
plot(lat$net, coord = lat$coord, jitter = FALSE, displaylabels = TRUE)
}
