% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_hexagon.R
\name{lattice_hexagon}
\alias{lattice_hexagon}
\title{Generate a Lattice for Hexagon}
\usage{
lattice_hexagon(nx = 8, ny = 8, torus = TRUE, display_plot = TRUE)
}
\arguments{
\item{nx}{Integer. Number of nodes along the x-axis. Must be 2 or greater.}

\item{ny}{Integer. Number of nodes along the y-axis. Must be 2 or greater.}

\item{torus}{Logical. If \code{TRUE}, the lattice wraps around horizontally and vertically to form a toroidal structure.}

\item{display_plot}{Logical. If \code{TRUE}, plot the resulting lattice using \code{plot.network()} from the \strong{network} package.}
}
\value{
A list with the following components:
\describe{
\item{\code{adj}}{A symmetric adjacency matrix representing the lattice for hexagon.}
\item{\code{coord}}{A data frame of node coordinates for plotting.}
\item{\code{net}}{A \code{network} object from the \strong{network} package representing the lattice.}
}
}
\description{
Construct an adjacency matrix for a hexagonal shape lattice with \eqn{n_x \times n_y} nodes, optionally wrapped as a torus.
}
\examples{
lat <- lattice_hexagon(nx = 8, ny = 8, torus = TRUE, display_plot = FALSE)
lat$adj
plot(lat$net, coord = lat$coord, jitter = FALSE, displaylabels = TRUE)
}
