% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_line.R
\name{lattice_line}
\alias{lattice_line}
\title{Generate a Lattice for Line}
\usage{
lattice_line(n = 5, display_plot = TRUE)
}
\arguments{
\item{n}{A positive integer specifying the number of nodes in the graph.}

\item{display_plot}{Logical. If \code{TRUE}, the lattice is plotted using \code{plot.network()} from the \strong{network} package.}
}
\value{
A list with the following components:
\describe{
\item{\code{adj}}{A symmetric adjacency matrix representing the lattice for line}
\item{\code{coord}}{A data frame of node coordinates for plotting.}
\item{\code{net}}{A \code{network} object from the \strong{network} package representing the lattice.}
}
}
\description{
Construct an adjacency matrix for a line (i.e., path graph) with \code{n} nodes.
}
\examples{
lat <- lattice_line(n = 5, display_plot = FALSE)
lat$adj
plot(lat$net, coord = lat$coord, jitter = FALSE, displaylabels = TRUE)
}
