% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_path.R
\name{lattice_path}
\alias{lattice_path}
\title{Generate a Path Graph}
\usage{
lattice_path(n = 5, display_plot = TRUE)
}
\arguments{
\item{n}{A positive integer specifying the number of nodes in the path graph.}

\item{display_plot}{Logical. If \code{TRUE}, the lattice is plotted using \code{plot.network()} from the \strong{network} package.}
}
\value{
A list with the following components:
\describe{
\item{\code{adj}}{A symmetric adjacency matrix representing the path graph.}
\item{\code{coord}}{A data frame of node coordinates arranged in a line.}
\item{\code{net}}{A \code{network} object from the \strong{network} package representing the graph.}
}
}
\description{
Construct an adjacency matrix for a path graph with \code{n} nodes.
}
\examples{
lat <- lattice_path(n = 5, display_plot = FALSE)
lat$adj
plot(lat$net, coord = lat$coord, jitter = FALSE, displaylabels = TRUE)
}
